RAG_Workbench/requirements.txt
# پردازش متن فارسی
hazm==0.7.0
python-docx==0.8.11

# Embedding و Vector DB
requests==2.31.0
chromadb==0.4.15
numpy

# UI
# tkinter is part of the standard library
Pillow==10.0.0

# Data Processing
# scikit-learn==1.3.0

# تست
pytest==7.4.0
pytest-cov==4.1.0